Update/fix/add tests for puppet-related flow (preassign+match) (#17859)
